+++ dart/lib/src/sentry_client.dart
@@ -211,14 +211,12 @@ class SentryClient {
 
   SentryEvent _prepareEvent(SentryEvent event, Hint hint,
       {dynamic stackTrace}) {
-    event = event.copyWith(
-      serverName: event.serverName ?? _options.serverName,
-      dist: event.dist ?? _options.dist,
-      environment: event.environment ?? _options.environment,
-      release: event.release ?? _options.release,
-      sdk: event.sdk ?? _options.sdk,
-      platform: event.platform ?? sdkPlatform(_options.platform.isWeb),
-    );
+    event.serverName = event.serverName ?? _options.serverName;
+    event.dist = event.dist ?? _options.dist;
+    event.environment = event.environment ?? _options.environment;
+    event.release = event.release ?? _options.release;
+    event.sdk = event.sdk ?? _options.sdk;
+    event.platform = event.platform ?? sdkPlatform(_options.platform.isWeb);
 
     if (event is SentryTransaction) {
       return event;
@@ -255,7 +253,7 @@ class SentryClient {
         if (!_options.platform.isWeb &&
             isolateName != null &&
             _options.attachThreads) {
-          sentryException = sentryException.copyWith(threadId: isolateId);
+          sentryException.threadId = isolateId;
           sentryThread = SentryThread(
             id: isolateId,
             name: isolateName,
@@ -270,13 +268,15 @@ class SentryClient {
         }
       }
 
-      return event.copyWith(
-        exceptions: [...?event.exceptions, ...sentryExceptions],
-        threads: [
-          ...?event.threads,
-          ...sentryThreads,
-        ],
-      );
+      event.exceptions = [
+        ...?event.exceptions,
+        ...sentryExceptions,
+      ];
+      event.threads = [
+        ...?event.threads,
+        ...sentryThreads,
+      ];
+      return event;
     }
 
     // The stacktrace is not part of an exception,
@@ -292,7 +292,7 @@ class SentryClient {
         removeSentryFrames: hint.get(TypeCheckHint.currentStackTrace),
       );
       if (sentryStackTrace.frames.isNotEmpty) {
-        event = event.copyWith(threads: [
+        event.threads = [
           ...?event.threads,
           SentryThread(
             name: isolateName,
@@ -301,7 +301,7 @@ class SentryClient {
             current: true,
             stacktrace: sentryStackTrace,
           ),
-        ]);
+        ];
       }
     }
 
@@ -309,16 +309,14 @@ class SentryClient {
   }
 
   SentryEvent _createUserOrSetDefaultIpAddress(SentryEvent event) {
-    final user = event.user;
+    var user = event.user;
     final effectiveIpAddress =
         user?.ipAddress ?? (_options.sendDefaultPii ? _defaultIpAddress : null);
 
     if (effectiveIpAddress != null) {
-      final updatedUser = user == null
-          ? SentryUser(ipAddress: effectiveIpAddress)
-          : user.copyWith(ipAddress: effectiveIpAddress);
-
-      return event.copyWith(user: updatedUser);
+      user ??= SentryUser(ipAddress: effectiveIpAddress);
+      user.ipAddress = effectiveIpAddress;
+      event.user = user;
     }
 
     return event;