+++ dart/test/protocol/sentry_browser_test.dart
@@ -36,7 +36,7 @@ void main() {
   group('copyWith', () {
     test('copyWith keeps unchanged', () {
       final data = sentryBrowser;
-
+      // ignore: deprecated_member_use_from_same_package
       final copy = data.copyWith();
 
       expect(
@@ -47,7 +47,7 @@ void main() {
 
     test('copyWith takes new values', () {
       final data = sentryBrowser;
-
+      // ignore: deprecated_member_use_from_same_package
       final copy = data.copyWith(
         name: 'name1',
         version: 'version1',